# Optimize RM_Call by recycling the client