This PR makes the client used by the module to perform RM_Call, part of the module context. If repeated calls to RM_Call are made, we reuse the same client, only destroying it along with the module.

This achieves a huge performance gain in repeated RM_Call's. For comparison, here is an example of **calling PING a million times** with and without this patch.

**Without:**

```
127.0.0.1:6379> BENCH.CALL
OKAY
(3.48s)
127.0.0.1:6379> BENCH.CALL
OKAY
(3.42s)
```

**After the patch is applied:**:
```
127.0.0.1:6379> BENCH.CALL
OKAY
(1.28s)
127.0.0.1:6379> BENCH.CALL
OKAY
(1.27s)
```

Notice that this only optimizes repeated calls within the same context, which is a rather common case but not always the case. If we just used the same client for ALL RM_Call's, we would optimize even for the case of a module performing a single call. I wasn't sure how safe that is, so I want with the more conservative option. If you think that we can recycle all the way - let's go for it!