@antirez 2018-01-22
That looks awesome! Creating the client is quite some work... and Btw Lua also reuses the same client again and again. I can't see how this is going to create a problem, at max you may forget to reset some flag in the clients you are reusing, but this can be fixed anyway retaining the performance gain, if we found a problem. I'll review the patch just to see if I can find any potential state in the client we want to reset, and merge anyway (either I add the commit to reset such state or merge as it is). Thanks