@antirez 2018-01-22
It's not related to this work, but at this point a good question could be, is client creation so expensive because there is no way to optimize this further, or we are doing too much work? For instance, does it make sense to have many data structures that many clients will never use to be initialized ASAP, or we should do that work on-demand?