@dvirsky 2018-01-22
@antirez also, can clients maybe be pooled and reused with a global pool? But reading the code it looks like a lot of this stuff is not really needed, especially by a module client, and can either be lazily created, made opt-in by some flag signaling or something, if we know for certain the client will not need it.