@antirez 2018-01-22
Yep, first step will be profiling the function to understand where all this time is burn. Certain things are obviously too aggressively done, for instance the list of Pub/Sub channels. The dictionary should be created on demand as soon as we need to add something to it, and so forth. Btw back on your patch, I'll review ASAP and merge so we get this benefit to start. At a first glance to reuse the client forever may also be possible, since there should never be concurrent access, but I need to do a serious analysis. For instance when we use threaded contexts we create a fake client just for the thread scope, so there is no danger of accessing the global client everybody is using. Similarly given that Lua Scripting is already "bug fixed" we can check how the shared client handling is performed there and what state is reset across calls. After all there are many module commands doing just 1 or 2 RM_Call() calls, so having a single shared client may improve performances in a very significant way. Of course the fact we have so poor testing of the modules API does not make changing things in a safe way trivial :-( The good thing however if that by testing if serious modules like RediSearch still work is somewhat a partial assurance.