@dvirsky 2018-01-22
yeah, I suppose if valgrind doesn't find anything when running the entire search test suite, we are safe to assume that we're not leaking or anything.