@concretevitamin 2018-01-22
@dvirsky this is awesome!

Re: your question earlier, my use case is basically the following
```c++
MyFancyCmd_RedisCommand() {
    ... // Some stuff, no calls.
    // Occurs once in this function.
    RedisModule_Call(ctx, "PUBLISH", channel, ...);
    ...
}
```
So there is only one `Call` in `MyFancyCmd`.  However, there are many (millions) of calls to `MyFancyCmd`, with only a few tens of distinct `channel`.  Is there an optimization opportunity here?

@antirez for this use case, any chance to expose a native, more efficient `RedisModule_Publish`?