@dvirsky 2018-01-22
@concretevitamin I've replied to your private email with a workaround involving thread safe contexts, but it's a pretty complex one...