@dvirsky 2018-01-22
FWIW I had a go at implementing a low level API for PUBLISH. https://github.com/antirez/redis/pull/4625