@concretevitamin 2018-01-23
If there's any development on `After all there are many module commands doing just 1 or 2 RM_Call() calls, so having a single shared client may improve performances in a very significant way.` -- which I'm very in favor of -- please update this thread :-)