@antirez 2018-01-24
Hello @concretevitamin, when I'll review this PR (before end of this week) I'll try to see if we can go with just a shared client for all the requests which is optimal. However while to optimize RM_Call() is a very worthwhile effort, in the long run we should end with low level APIs for almost anything... Like RM_Publish() PRed by @dvirsky. More info ASAP.