@concretevitamin 2018-01-24
@antirez Thanks. I agree with most of your points, with one counter(?)example.  Although we already have the low-level `RM_StringSet`, I've found it to have significant overheads due to `RM_OpenKey` and `RM_CloseKey`, which consist of ~70% of the actual StringSet runtime.  

This is why I'd really like to see a shared client for all Call's, to benchmark against `RM_Call("SET")`.