@oranagra 2021-02-07
i think merging it as is would be missing the opportunity to do something better.
i think this is a great improvement but we do need to make sure it applies for modules that do only one call per context.
so the two possible ways to improve it are:
1. find out what's so heavy client creation and improve it for the benefit of everyone, not just modules
2. create a cached client like the one used for Lua, that is used for all RM_Call (on behalf of different modules)
  - we'll need to reset some of the client fields and flags after each use (possibly reusing some of the logic of the RESET command)
  - we'll need to make sure that calls not made from within the main command execution pipeline (redis's main thread) are not using this cached client)

maybe @filipecosta90 wanna look into the profiling part and / or @MeirShpilraien at the other part (considering you are a heavy user of these)