@MeirShpilraien 2021-02-07
@oranagra the problem I see with one singe client for all rm_call (like Lua) is recursive rm_call (module that call anothe module..). So I think we need a pool of clients if we want to do it correctly... (or stay with the current optimization suggested in this PR)

Another question, I am not sure about your last point, why we can not use the same shared client/client pool from background thread? When we do RM_Call we take the GIL so I do not see the problem, what am I missing?