@oranagra 2021-02-07
@MeirShpilraien i think i would be ok with making an optimization just for the first level of nesting, and have the second level create a client each time, or create one on the first call, and destroy it when exiting (i.e. when RM_Call returns it destroys the second second level if there is one. possibly with a stack of clients).
i think merging the optimization like it is would miss many cases of single call to RM_Call, which are more common than nesting so i rather optimize that and leave the other, than not optimize either.

regarding background thread, i guess you're right. i thought about concurrent use of a client, but that can't happen since obviously the GIL will only be released by the main thread when the client isn't used.